Add test for has_ingoing_edges if the DiskPath storage
