+++ graphannis/Cargo.toml
@@ -52,7 +52,7 @@ transient-btree-index = "0.5"
 zip = "0.6.4"
 
 [dev-dependencies]
-criterion = "0.4"
+criterion = "0.5"
 fake = "2.2"
 insta = "1.34.0"
 pretty_assertions = "1.3"
